Move devices to models.py
