--- homeassistant/components/walkingpad/translations/ru.json
@@ -1,41 +1,0 @@
-{
-  "config": {
-    "abort": {
-      "already_configured": "Устройство уже настроено.",
-      "no_devices_found": "Устройства не обнаружены. Попробуйте перезагрузить WalkingPad.",
-      "ble_not_available": "Bluetooth не доступен."
-    },
-    "error": {
-      "cannot_connect": "Подключение не удалось."
-    },
-    "step": {
-      "user": {
-        "description": "Выберите тип подключения. Китайские версии WalkingPad обычно подключаются с помощью WiFi, европейские - с помощью Bluetooth.",
-        "data": {
-          "conn_type": "Тип подключения."
-        }
-      },
-      "ble_device": {
-        "data": {
-          "device": "Устройство"
-        }
-      },
-      "wifi_device": {
-        "title": "IP адрес и токен.",
-        "data": {
-          "ip_address": "IP адрес",
-          "token": "Токен"
-        }
-      }
-    }
-  },
-  "options": {
-    "step": {
-      "init": {
-        "data": {
-          "default_speed": "Скорость по умолчанию"
-        }
-      }
-    }
-  }
-}
